Extracted FlowsList into own file
