+++ rasa/dialogue_understanding/commands/can_not_handle_command.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from typing import Any, Dict, List
 from rasa.dialogue_understanding.commands import Command
 from rasa.shared.core.events import Event
-from rasa.shared.core.flows.flow import FlowsList
+from rasa.shared.core.flows.flows_list import FlowsList
 from rasa.shared.core.trackers import DialogueStateTracker
