+++ rasa/dialogue_understanding/commands/knowledge_answer_command.py
@@ -6,7 +6,7 @@ from rasa.dialogue_understanding.commands import FreeFormAnswerCommand
 from rasa.dialogue_understanding.patterns.search import SearchPatternFlowStackFrame
 from rasa.dialogue_understanding.stack.dialogue_stack import DialogueStack
 from rasa.shared.core.events import Event
-from rasa.shared.core.flows.flow import FlowsList
+from rasa.shared.core.flows.flows_list import FlowsList
 from rasa.shared.core.trackers import DialogueStateTracker
