+++ tests/dialogue_understanding/commands/conftest.py
@@ -3,7 +3,7 @@ import pytest
 from rasa.dialogue_understanding.commands import StartFlowCommand
 from rasa.dialogue_understanding.processor.command_processor import execute_commands
 from rasa.shared.core.events import UserUttered
-from rasa.shared.core.flows.flow import FlowsList
+from rasa.shared.core.flows.flows_list import FlowsList
 from rasa.shared.core.trackers import DialogueStateTracker
 from rasa.shared.nlu.constants import COMMANDS
 from rasa.shared.core.flows.yaml_flows_io import flows_from_str
