+++ tests/dialogue_understanding/commands/test_command_processor.py
@@ -11,7 +11,7 @@ from rasa.dialogue_understanding.stack.frames import (
     PatternFlowStackFrame,
 )
 from rasa.shared.core.constants import FLOW_HASHES_SLOT
-from rasa.shared.core.flows.flow import FlowsList
+from rasa.shared.core.flows.flows_list import FlowsList
 from rasa.shared.core.trackers import DialogueStateTracker
 from rasa.shared.core.flows.yaml_flows_io import flows_from_str
 from tests.dialogue_understanding.commands.conftest import start_bar_user_uttered
