+++ tests/dialogue_understanding/commands/test_handle_code_change_command.py
@@ -17,11 +17,11 @@ from rasa.dialogue_understanding.stack.frames import (
 from rasa.shared.core.domain import Domain
 from rasa.shared.core.events import SlotSet
 from rasa.shared.core.flows.flow import (
-    FlowsList,
     START_STEP,
     ContinueFlowStep,
     END_STEP,
 )
+from rasa.shared.core.flows.flows_list import FlowsList
 from rasa.shared.core.trackers import DialogueStateTracker
 from tests.dialogue_understanding.commands.test_command_processor import (
     start_bar_user_uttered,
